[Sim] Fix WeightsMutator and Stadium
Summary:
WeightsMutator setting was out of order. Stadium stat should always be Stadium.. if it's
home, the data in the 'Stadium' field should just be the home data. I.e. this should be done
before the data gets to the sim.

Test Plan:
Run test.py, spits out WeightsMutator overridden data when uncommented, spits out
originally set data when not.

Reviewers: cert

Reviewed By: cert

Differential Revision: http://phabricator.sabertoothventures.com/D40