Create mem_exe_sql function to check memcached before reads
Summary:
mem_exe_sql function works the same as exe_sql but checks the cache
before the sql call. If something is not in the cache it works like exe_sql
but then adds query and result to the cache afterwards -> defaulted to 1hr
retention at the moment we can revisit later.

Test Plan:
Ran a similar version on your machine but will have to test after
you accept since I don't have memcached (or good data) on my computer.

Reviewers: smas

Reviewed By: smas

Differential Revision: http://phabricator.sabertoothventures.com/D114