Modify Pitcher Parse For Previous/Career Pitcher Pull
Summary:
Only pull reliever data for previous and career runs for known relief pitchers for the current season.
Wasn't coding at my finest last night but even looking at it this morning not sure if there is a more elegant way
to accomplish this.

Test Plan: ran on bb comp

Reviewers: smas

Reviewed By: smas

Differential Revision: http://phabricator.sabertoothventures.com/D178