Redo Logging For Table Status
Summary: Before logging was through multi-insert - since this gets called several times in the scripts (for seach split for instances in espn_batting, etc.) if the script does fail we don't see it. Instead making it a seperate function call at the end of each script.

Test Plan: none - testing in prod

Reviewers: smas

Reviewed By: smas

Differential Revision: http://phabricator.sabertoothventures.com/D207