Add better error handling for the lineups script
Summary: Before the script just returned null if a player_id could not be found. This means our sim won't have all the best data cause of typos and is not what we want. This was also me just playing around with error handling and is likely not the final solution. I need to decide if I like it (a) emailing, (b) still running even when the exception is thrown, etc. but for now I wanted something better than just returning null and not sending us any signal.

Test Plan:
ran and got email below:
{F1881}

Reviewers: smas

Reviewed By: smas

Differential Revision: http://phabricator.sabertoothventures.com/D213