Create DTs For Live Scores & Odds
Summary: Data types, plus modify THasSimParams to have getters. Create setColumns function for datatypes for dealing with DT's (like bets) where we have two different use cases. Also created a NotParams function for DTs to support <> and isNotNull.

Test Plan: Will have to test the new DT's in prod, but created test of the private function that creates the where statements since I made that logic a bit more complex. I know you're supposed to test the public function but it was much easier to verify the sql than to create fake data and compare the objects. This will likely add more columns to the game page but I can fix that in another diff.

Reviewers: smas

Reviewed By: smas

Differential Revision: http://phabricator.sabertoothventures.com/D226