Create Tables::getColheads function and Add Basic DailyScriptWithWrite Class
Summary: Breaking down the larger diff into smaller pieces. The ScriptWWrites class will be the base of most daily things with the ability to cycle through days/backfill as well as have a test field. I modified the multi_insert function to accept null colheads assuming they are set in the Tables constants class. Also made a daily include class.

Test Plan: Tested calling multi_insert with the Bets table (to return the colheads) and then once with an unfilled table to confirm the exception was thrown.

Reviewers: smas

Reviewed By: smas

Differential Revision: http://phabricator.sabertoothventures.com/D246