Create new MySQL class with insert function (so far)
Summary: (a) do we want to keep having to write MySQL? Should be be lower case? All uppercase? This diff is just the insert() and I will add unit tests for these things in a seperate diff. Let me know if I went too exception heavy :) - also added function to ArrayUtils for array of arrays.

Test Plan: Ran test.php will keep around for the other functions in this class and delete later.

Reviewers: smas

Reviewed By: smas

Differential Revision: http://phabricator.sabertoothventures.com/D268