Add Try Catch On Live Odds Updater
Summary: So this doesn't update if odds haven't moved so this error is throwing a lot. I can't decide if catching it is a bad idea? Thoughts? It would seem if there were no odds the error could come up in the Bets script and that'd be the useful error?

Test Plan: none

Reviewers: smas

Reviewed By: smas

Differential Revision: http://phabricator.sabertoothventures.com/D310